Move e2e tests to where they belong
